northflank/requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
numpy==2.4.1
snowflake-connector-python==3.12.3
